Fixed up a couple naming issues such as camelCase vs PascalCase and ran through the whole game and it all appears to be working just fine. Ill continue to look through the code this afternoon and then have an instructor review it first thing tomorrow

+++ LemonadeStand/UserInterface.cs
@@ -75,7 +75,7 @@ namespace LemonadeStand
             Console.WriteLine("Error. Please input again.\n");
         }
 
-        public string StorePrices(double[] StorePrices)
+        public string StorePrices(int[] StorePrices)
         {
             Console.ForegroundColor = ConsoleColor.DarkCyan;
             Console.Write("The current store prices are: " + StorePrices[0] + " grubs per banana, ");
